Fix classifier training crash on DeBERTa-v3: use bf16, not fp16
fp16 mixed precision with DeBERTa-v3's disentangled attention hit
"Attempting to unscale FP16 gradients" during backprop (RoBERTa
trained fine under the same code — this is DeBERTa-v3 specific).
bf16 doesn't use loss scaling at all, so the failure mode doesn't
exist, and Ampere+ GPUs (RTX 4090 included) support it natively.

+++ textscope/classifier.py
@@ -253,7 +253,13 @@ def train_classifier(
         logging_steps=50,
         learning_rate=2e-5,
         weight_decay=0.01,
-        fp16=(device == "cuda"),
+        # bf16, not fp16: DeBERTa-v3's disentangled-attention relative
+        # position embeddings hit "Attempting to unscale FP16 gradients"
+        # under fp16's loss-scaling GradScaler on some parameter shapes.
+        # bf16 has fp32's exponent range, so it needs no loss scaling and
+        # sidesteps the bug entirely; every CUDA GPU new enough to be
+        # running this (Ampere+) supports it natively.
+        bf16=(device == "cuda"),
         report_to=[],
     )
